+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/charts/__tests__/ChartWithExport.test.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import { cleanup } from '@testing-library/react';
 
-import ChartWithExport from '../ChartWithExport';
 import { renderWithProviders } from '../../../../setupTests';
+import { ChartWithExport } from '../chart-with-export';
 
 afterEach(cleanup);