+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/charts/chart-with-export/index.js
@@ -1,0 +1,1 @@
+export { default as ChartWithExport } from './ChartWithExport';