+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/charts/points-distribution-chart/index.js
@@ -1,0 +1,1 @@
+export { default as PointsDistributionChart } from './PointsDistributionChart';